requirements.txt
PyPDF2==1.26.0
blinker==1.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
defusedxml==0.5.0
django-wkhtmltopdf==3.2.0
docxtpl==0.5.6
Flask==1.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.0
pdfkit==0.6.1
psycopg2==2.7.5
psycopg2-binary==2.7.6.1
PyJWT==1.6.4
python-docx==0.8.7
python3-openid==3.1.0
requests==2.21.0
six==1.11.0
SQLAlchemy==1.2.12
urllib3==1.24.1
Werkzeug==0.14.1
wkhtmltopdf==0.2
wkhtmltopdf-pack-ng==0.12.3.0
WTForms==2.2.1
xlrd==1.1.0
